+++ atoum/views/__init__.py
@@ -1,14 +1,22 @@
-from .assortment import AssortmentAutocompleteView
-from .category import CategoryAutocompleteView
-from .consumable import RecursiveTreeView
+from .assortment import AssortmentAutocompleteView, AssortmentDetailView
+from .category import CategoryAutocompleteView, CategoryDetailView
+from .consumable import ConsumableDetailView, ConsumableIndexView
+from .dashboard import DashboardView
 from .dummy import DummyView
-from .product import ProductAutocompleteView
+from .product import ProductAutocompleteView, ProductDetailView
+from .tree import RecursiveTreeView
 
 
 __all__ = [
     "AssortmentAutocompleteView",
+    "AssortmentDetailView",
     "CategoryAutocompleteView",
+    "CategoryDetailView",
+    "ConsumableDetailView",
+    "ConsumableIndexView",
+    "DashboardView",
     "DummyView",
     "ProductAutocompleteView",
+    "ProductDetailView",
     "RecursiveTreeView",
 ]